PT-6: Refactor sender & add tests
