+++ src/VirtoCommerce.StoreModule.Data/Services/StoreNotificationSender.cs
@@ -10,9 +10,9 @@ using VirtoCommerce.NotificationsModule.Core.Types;
 using VirtoCommerce.Platform.Core.Common;
 using VirtoCommerce.Platform.Core.Security;
 using VirtoCommerce.Platform.Core.Settings;
-using VirtoCommerce.StoreModule.Core;
 using VirtoCommerce.StoreModule.Core.Model;
 using VirtoCommerce.StoreModule.Core.Services;
+using StoreSettings = VirtoCommerce.StoreModule.Core.ModuleConstants.Settings;
 
 namespace VirtoCommerce.StoreModule.Data.Services
 {
@@ -23,7 +23,6 @@ namespace VirtoCommerce.StoreModule.Data.Services
         private readonly IStoreService _storeService;
         private readonly Func<UserManager<ApplicationUser>> _userManagerFactory;
 
-
         public StoreNotificationSender(INotificationSearchService notificationSearchService, INotificationSender notificationSender, IStoreService storeService, Func<UserManager<ApplicationUser>> userManagerFactory)
         {
             _notificationSearchService = notificationSearchService;
@@ -32,46 +31,36 @@ namespace VirtoCommerce.StoreModule.Data.Services
             _userManagerFactory = userManagerFactory;
         }
 
-
         public virtual async Task<NotificationSendResult> SendUserEmailVerificationAsync(ApplicationUser user)
         {
-            var result = new NotificationSendResult();
-
+            // Return error if store not found
             var store = await _storeService.GetByIdAsync(user.StoreId, StoreResponseGroup.None.ToString());
-
-            if (store != null)
+            if (store is null) return new NotificationSendResult
             {
-                var isVerificationEnabled = store.Settings.GetSettingValue(ModuleConstants.Settings.General.EmailVerificationEnabled.Name,
-                                                                     (bool)ModuleConstants.Settings.General.EmailVerificationEnabled.DefaultValue);
-                if (isVerificationEnabled)
-                {
-                    var notification = await _notificationSearchService.GetNotificationAsync<ConfirmationEmailNotification>(new TenantIdentity(store.Id, nameof(Store)));
-                    if (notification != null)
-                    {
-                        notification.Url = await GenerateEmailVerificationLink(user, store);
-                        notification.From = store.AdminEmailWithName ?? store.EmailWithName;
-                        notification.To = user.Email;
+                ErrorMessage = $"Can't find store for user {user.UserName}."
+            };
 
-                        result = await _notificationSender.SendNotificationAsync(notification);
-                    }
-                    else
-                    {
-                        result.ErrorMessage = $"Can't find {nameof(ConfirmationEmailNotification)} notification for store {store.Id}.";
-                    }
-                }
-                else
-                {
-                    result.ErrorMessage = $"Email verification sending disabled. Check {ModuleConstants.Settings.General.EmailVerificationEnabled.Name} setting for store {store.Id}.";
-                }
-            }
-            else
+            // Return error if email verification feature is disabled
+            var settingDescriptor = StoreSettings.General.EmailVerificationEnabled;
+            var isVerificationEnabled = store.Settings.GetSettingValue(settingDescriptor.Name, (bool)settingDescriptor.DefaultValue);
+            if (!isVerificationEnabled) return new NotificationSendResult
             {
-                result.ErrorMessage = $"Can't find store for user {user.UserName}.";
-            }
+                ErrorMessage = $"Email verification sending disabled. Check {settingDescriptor.Name} setting for store {store.Id}."
+            };
 
-            return result;
+            // Return error if notification is not registered
+            var notification = await _notificationSearchService.GetNotificationAsync<ConfirmationEmailNotification>(new TenantIdentity(store.Id, nameof(Store)));
+            if (notification is null) return new NotificationSendResult
+            {
+                ErrorMessage = $"Can't find {nameof(ConfirmationEmailNotification)} notification for store {store.Id}."
+            };
+
+            notification.Url = await GenerateEmailVerificationLink(user, store);
+            notification.From = store.AdminEmailWithName ?? store.EmailWithName;
+            notification.To = user.Email;
+
+            return await _notificationSender.SendNotificationAsync(notification);
         }
-
 
         public virtual async Task<string> GenerateEmailVerificationLink(ApplicationUser user, Store store)
         {
@@ -80,18 +69,15 @@ namespace VirtoCommerce.StoreModule.Data.Services
                 using var userManager = _userManagerFactory();
                 var token = await userManager.GenerateEmailConfirmationTokenAsync(user);
 
-                var result = QueryHelpers.AddQueryString(new Uri($"{store.Url.TrimEnd('/')}/account/confirmemail").ToString(),
+                return QueryHelpers.AddQueryString(new Uri($"{store.Url.TrimEnd('/')}/account/confirmemail").ToString(),
                      new Dictionary<string, string>
                      {
                         { "UserId", user.Id },
                         { "Token", token }
                     });
-                return result;
             }
-            else
-            {
-                throw new OperationCanceledException($"A valid URL is required in Url property for store '{store.Id}'.");
-            }
+
+            throw new OperationCanceledException($"A valid URL is required in Url property for store '{store.Id}'.");
         }
     }
 }
